requirements.txt
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0
slack-sdk==3.26.1
flask==3.0.0
anthropic==0.39.0
